Create new pattern for AST and refactor parser
Uses new AST which will replace old one, stores items which represent
the things being executed (e.g. the list of top level statements, but
may include things in the future like class definitions).

Statements and individual expressions are stored in a global list and
their IDs are used in reference when building up other
statements/expressions. This makes the tree have a more indirect
structuring and lets individual statements and expressions (and
sub-expressions) be referenced and modified in the AST.

Remaining work:
- Add visitor pattern
- Implement visitor pattern for evaluator
- Complete refactoring parser
- Utilize new AST in main
- Delete old code