Add parsing for semicolons
All expression statements in a block must end with a semicolon unless
they are the last one. This is enforced by the type checker.

This required defining two different ways to do if, as expressions and
as statments. This allows if statements to not require a semicolon, but
when used as an expression the parent kind can decide if it needs one or
not (e.g. let expressions must end with a semicolon).