Add better type handling
Includes setting a type for the return stmt (mainly just makes it easier
to check later on, since you could always just query the expr and get
its type instead).
Also adds a wrapper around Vec<Type> to make it easier to reproduce
error messages for lists of types (e.g. when calling or when printing
out a function signature)